+++ component-utils/documentation/js/__config/polyfills.js
@@ -1,0 +1,4 @@
+export default {
+  'fetch': () => import('isomorphic-fetch'),
+  'Object.assign': () => import('@aofl/polyfill-service/src/object-assign-polyfill')
+};