+++ component-utils/documentation/routes/home-element/find-parent-examples/template.js
@@ -1,0 +1,5 @@
+(context, html) => html`
+  <parent-comp>
+    <child-comp></child-comp>
+  </parent-comp>
+`;