+++ component-utils/index.js
@@ -1,0 +1,5 @@
+import findParent from './src/find-parent';
+
+export {
+  findParent
+};